clustering code cleanup

+++ cz.mzk.recordmanager.server/src/main/java/cz/mzk/recordmanager/server/util/DeduplicationUtils.java
@@ -6,11 +6,11 @@ public class DeduplicationUtils {
 	 * 
 	 * @param pages1
 	 * @param pages2
-	 * @param boundary maximal percentual difference
+	 * @param boundary maximal percentual difference, value [0-100]
 	 * @param absoluteDiference maximal absolute difference
 	 * @return true if false if pages, true if at least one is null or they are in safe interval
 	 */
-	public static boolean comparePages(Long pages1, Long pages2, double boundary, int absoluteDiference) {
+	public static boolean comparePages(Long pages1, Long pages2, int boundary, int absoluteDiference) {
 		if (pages1 == null || pages2 == null) {
 			return true;
 		}
